Organize buffers tests
